// apps/backend/src/dto/supplier.dto.ts
export type SupplierDTO = {
  id?: number;
  name: string;
  cnpj?: string;
  email?: string;
  phone?: string;
  cep?: string;
  address?: string;
  number?: string;
  neighborhood?: string;
  city?: string;
  state?: string;
  complement?: string;
  created_at: string;
  active?: boolean;
  observation?: string;
};